Add support for closeable bundle filters.
Great effort is made to close these filters after each request.
The hydra tutor has some support for filter caching. This allows
for the testing of stateful filters, although those filters are in
the minority. op:"count" is one example. The cache is wiped clean
any time a closeable bundle filter is used.